+++ pbp/main_hmb_generator.py
@@ -1,13 +1,13 @@
+from argparse import Namespace
+
 from pbp.main_hmb_generator_args import parse_arguments
 
 # Some imports, in particular involving data processing, cause a delay that is
 # noticeable when just running the --help option. We get around this issue by
-# postponing the imports until actually needed. See the main() function.
+# postponing the imports until actually needed.
 
 
-def main():
-    opts = parse_arguments()
-
+def run_main_hmb_generator(opts: Namespace) -> None:
     # pylint: disable=import-outside-toplevel
     import os
 
@@ -66,6 +66,8 @@ def main():
         file_helper=file_helper,
         output_dir=opts.output_dir,
         output_prefix=opts.output_prefix,
+        compress_netcdf=opts.compress_netcdf,
+        add_quality_flag=opts.add_quality_flag,
         global_attrs_uri=opts.global_attrs,
         set_global_attrs=opts.set_global_attrs,
         variable_attrs_uri=opts.variable_attrs,
@@ -82,4 +84,4 @@ def main():
 
 
 if __name__ == "__main__":
-    main()
+    run_main_hmb_generator(parse_arguments())