+++ pbp/main_hmb_generator_args.py
@@ -136,6 +136,22 @@ Examples:
     )
 
     parser.add_argument(
+        "--no-netcdf-compression",
+        dest="compress_netcdf",
+        default=True,
+        action="store_false",
+        help="Do not compress the generated NetCDF file.",
+    )
+
+    parser.add_argument(
+        "--add-quality-flag",
+        dest="add_quality_flag",
+        default=False,
+        action="store_true",
+        help="Add quality flag variable (with value 2 - 'Not evaluated') to the generated NetCDF file.",
+    )
+
+    parser.add_argument(
         "--s3",
         default=False,
         action="store_true",